switch case to respond to different commands
